+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Nested-field returning changed line index
+- Nested-field enumerating label
+
 ## 2.5.1 - 2021-07-23
 
 ### Fixed