+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed 
+
+- Storybook
+- Lint
+- Postcss
+
 ## 2.5.1 - 2021-07-23
 
 ### Fixed